Move first conditional AI seeder to Ai namespace
